requirements.txt
APScheduler==3.0.2
CherryPy==3.7.0
Flask==0.10.1
Flask-Bower==1.1.1
git+https://github.com/youtux/flask-oauthlib.git@1454a6bcbb5248715363b416317a8c63315cfe89#egg=Flask-OAuthlib
Flask-SSLify==0.1.5
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
oauthlib==0.7.2
pymongo==3.0.1
pypebbleapi==0.0.1
pytz==2015.2
rauth==0.7.1
requests==2.6.2
requests-futures==0.9.5
requests-oauthlib==0.4.2
six==1.9.0
tzlocal==1.1.3
Werkzeug==0.10.4
Pillow==2.9.0
